Refactor design.

+++ HealthInMind/app/src/main/java/course/labs/healthinmind/medecine/data/abstractions/MedicineRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface MedicineRepository {
 
-    Medicine getMedicine(int id);
+    Medicine getMedicine(long id);
     Long createMedicine(Medicine medicine);
     Medicine updateMedicine(Medicine medicine);
     void deleteMedicine(int id);
