+++ HealthInMind/app/src/main/java/course/labs/healthinmind/medecine/data/abstractions/MedicinesLocalProvider.java
@@ -3,7 +3,7 @@ package course.labs.healthinmind.medecine.data.abstractions;
 import java.util.List;
 
 public interface MedicinesLocalProvider {
-    Medicine getMedicine(int id);
+    Medicine getMedicine(long id);
     Long createMedicine(Medicine medicine);
     Medicine updateMedicine(Medicine medicine);
     void deleteMedicine(int id);
